ride_sp.pl. TE colour in genbank output.
The coordinates for the TE reported by rodeo are off by one on one
end. Since exact coordinate match was being used for imparting a
colour to the TE, it was failing. Now I tolerate a mismatch of two
nucleotide positions. I am not very happy with this but cannot
think of any other solution. See patch below.

@@ -474,7 +474,9 @@ for my $subft (@subft) {
     unless ($protlen >= $minPPlen and $protlen <= $maxPPlen) {
       push(@recoord, [$st, $en, $str]);
     }
-    if($st == $teSubStart and $en == $teSubEnd) {
+    # if($st == $teSubStart and $en == $teSubEnd) {
+    if(abs($st - $teSubStart) <= 2 and
+      abs($en - $teSubEnd) <= 2) {
       $subft->add_tag_value("color", "0 255 0");
     }
   }