ride_sp.pl: Work around TE not in rodeo output.
If TE coordinates cannot be picked up from main_co_occur.csv
then, instead of giving up, we download the genbank file reported
in main_co_occur.csv and try to find the protein id inside this
file. If we can find it we get the coordinates from there and
push them to @coords after which we can carry on as usual. The
subroutine which does this is named TEcoordsByProtId. It takes a
has with two keys; gbkfile and protid.

Other scripts which have changed have simply had one comment
removed to reduce noise on the terminal.