"Colour" to "Color in make_cytoscape_attribute_file.pl
In response to Alaster's email pasted below.

Hi All,
Thanks for this Andy, it's been a long time coming.
I've had a scan through the "Delta PDF" file, and I can't see anything
significant beyond the following (see attached screenshot):
The network can now be coloured by using the “‘Style”’ pane. Under the
“‘Fill Color”’ option, select “‘Column: Colour’,”
They have corrected Column: Colour to Column: Color. While I understand
their reasons for doing so in the rest of the manuscript, in this
instance, it is actually spelled "colour" under the dropdown menu in
Cytoscape, as the RiPPER-EGN workflow has actually named the column
"colour". So, we can either make that edit, or see if Govind would
quickly change the text in the original code for RiPPER-EGN?
best,
Alaster