mkcooc.pl edited to do the entire list of accessions.
Earlier we used to process each input accession with mkcooc.pl
followed by ripper.pl. Now the idea is to process all the input
accessions with mkcooc.pl and then come to ripper.pl. This should
allow us to focus on fetching from Genbank in mkcooc.pl and not do any
fetching in ripper.pl. So once we get to ripper.pl we know that we
have everything needed to proceed. This version of mkcooc.pl works in
a standalone test in /temp/norod/ on n108524. ripper.pl will not be
able to work with the output produced by this version of mkcooc.pl.
Next we should edit ripper.pl but, mkcooc.pl will also continue to
change for some time.