Writing ripper_run.pl to replace ripper_run.sh.
Primarily this is to make the handling of command line arguments easier
for my self as well as to provide users with more command line options
to tweak the run of ripper.pl. At this stage it compiles and runs but
does not actually run any of the analysis scripts.