Use exit code from system call to mkcooc.pl.
If mkcooc.pl fails to get a nucleotide genbank file for a protein id
then it die()s rather than exit normally. This sets the exit code to
something other than zero. This exit code is used in ripper_run.pl to
decide whether to run ripper.pl or not.