HMM searches on prepeptides changed.
In pfam_sqlite.pl both collections (ripp and pfam) are searched but now
the name of the collection is included in the record of the hit.

In mergeRidePfam.pl for the distant prepeptides only those HMM hits are
reported which come from the ripp collection. For the distant
prepeptides all hits are reported.